gl: fix build for latest glm update
We had compiler errors like this:

  src/lib/di/algorithms/RenderIllustrativeLines.cpp:507:88: error: call of overloaded ‘setUniform(const char [9], glm::dvec3)’ is ambiguous
               m_composeShaderProgram->setUniform( "u_bbSize", getBoundingBox().getSize() );
                                                                                          ^
  src/lib/di/algorithms/RenderIllustrativeLines.cpp:507:88: note: candidates are:
               void setUniform( const std::string name, const glm::vec3& value )
               void setUniform( const std::string name, const glm::vec2& value )
                    ^

Probably nowadays glm-0.9.8.4 can cast glm::dvec3 to glm::vec3 and also to glm::vec2. That's why
I hate C++ and all this overloading ...

We fix it here by adding functions for dvec3 and friends. Intead of this
we could have added an expilcit cast to from dvec3 to vec3 to not change
the current behavior. Not sure what is better.